Update Migrate-Printers
Removed old Port deletion (REM'd out) as cannot successfully delete a PAPERCUT port.  Added query to dhcp server by IP address in order to get the attached Netbios name (if it exists) - The NBName is then used to check for existence in DNS and delete it if it's there in preparation for the new one.